refactor: ディレクトリ構成を修正
