+++ app/src/shared/errors/FetchError.ts
